@@ -2,6 +2,7 @@
 
 namespace Codewiser\Workflow;
 
+use Closure;
 use Codewiser\Workflow\Exceptions\TransitionException;
 use Codewiser\Workflow\Exceptions\TransitionFatalException;
 use Codewiser\Workflow\Exceptions\TransitionRecoverableException;
@@ -15,80 +16,97 @@ use Illuminate\Support\Str;
 /**
  * Transition between states in State Machine
  * @package Codewiser\Workflow
+ *
  */
 class Transition implements Arrayable
 {
-    protected $source;
-    protected $target;
+    protected string $source;
+    protected string $target;
     /**
      * @var Model|Workflow
      */
-    protected $model;
-    /**
-     * @var string
-     */
-    protected $attribute;
-    /**
-     * @var Collection|callable[]
-     */
-    protected $conditions;
-    /**
-     * @var Collection|callable[]
-     */
-    protected $callbacks;
+    protected Model $model;
+    protected string $attribute;
+    protected Collection $conditions;
+    protected Collection $callbacks;
+    protected Collection $attributes;
+    protected ?string $ability;
 
     /**
-     * These attributes must be provided into transit() method
-     * @var Collection
-     */
-    protected $attributes;
-
-    /**
-     * Instantiate new transition
+     * Instantiate new transition.
+     *
      * @param $source
      * @param $target
      * @return static
      */
-    public static function define($source, $target)
+    public static function define($source, $target): Transition
     {
         return new static($source, $target);
     }
 
-    public function __construct($source, $target)
+    public function __construct(string $source, string $target)
     {
         $this->source = $source;
         $this->target = $target;
+        $this->ability = null;
         $this->conditions = new Collection();
         $this->attributes = new Collection();
         $this->callbacks = new Collection();
     }
 
     /**
-     * Add condition to the transition
-     * @param callable $condition
-     * @return static
+     * Vivify transition with model context.
+     *
+     * @param Model $model
+     * @param string $attribute
      */
-    public function condition($condition)
+    public function inject(Model $model, string $attribute)
+    {
+        $this->model = $model;
+        $this->attribute = $attribute;
+    }
+
+    /**
+     * Authorize transition using policy ability.
+     * 
+     * @param string $ability
+     * @return $this
+     */
+    public function authorize($ability)
+    {
+        $this->ability = $ability;
+        return $this;
+    }
+
+    /**
+     * Add condition to the transition.
+     *
+     * @param Closure $condition
+     * @return $this
+     */
+    public function condition(Closure $condition)
     {
         $this->conditions->push($condition);
         return $this;
     }
 
     /**
-     * Callback(s) will run after transition is done
-     * @param callable $callback
+     * Callback(s) will run after transition is done.
+     *
+     * @param Closure $callback
      * @return $this
      */
-    public function callback($callback)
+    public function callback(Closure $callback)
     {
         $this->callbacks->push($callback);
         return $this;
     }
 
     /**
-     * Add requirement(s) to transition payload
+     * Add requirement(s) to transition payload.
+     *
      * @param string|string[] $attributes
-     * @return static
+     * @return $this
      */
     public function requires($attributes)
     {
@@ -114,7 +132,8 @@ class Transition implements Arrayable
     }
 
     /**
-     * Get human readable transition caption
+     * Get human readable transition caption.
+     *
      * @param bool $pastPerfect get caption for completed transition
      * @return array|Translator|string|null
      */
@@ -124,26 +143,39 @@ class Transition implements Arrayable
     }
 
     /**
-     * Source state
+     * Source state.
+     *
      * @return string
      */
-    public function getSource()
+    public function getSource(): string
     {
         return $this->source;
     }
 
     /**
-     * Target state
+     * Target state.
+     *
      * @return string
      */
-    public function getTarget()
+    public function getTarget(): string
     {
         return $this->target;
     }
 
     /**
-     * Get registered preconditions
-     * @return Collection|callable[]
+     * Ability to authorize.
+     * 
+     * @return string
+     */
+    public function getAbility(): ?string
+    {
+        return $this->ability;
+    }
+
+    /**
+     * Get registered preconditions.
+     *
+     * @return Collection|Closure[]
      */
     public function getConditions()
     {
@@ -151,63 +183,67 @@ class Transition implements Arrayable
     }
 
     /**
-     * Get registered transition callbacks
-     * @return callable[]|Collection
+     * Get registered transition callbacks.
+     *
+     * @return Collection|Closure[]
      */
     public function getCallbacks()
     {
         return $this->callbacks;
     }
 
-    public function __call($name, $arguments)
-    {
-        switch ($name) {
-            case 'inject':
-                $this->model = $arguments[0];
-                $this->attribute = $arguments[1];
-                break;
-        }
-    }
-
     /**
-     * Get list of problems with the transition
+     * Get list of problems with the transition.
+     *
      * @return array|string[]
      */
-    public function getProblems()
+    public function getProblems(): array
     {
-        $problems = [];
-        foreach ($this->getConditions() as $condition) {
-            try {
-                call_user_func($condition, $this->model);
-            } catch (TransitionFatalException $e) {
-                continue;
-            } catch (TransitionRecoverableException $e) {
-                $problems[] = $e->getMessage();
-            }
-        }
-        return $problems;
+        return $this->getConditions()
+            ->filter(function(\Closure $condition) {
+                try {
+                    call_user_func($condition, $this->model);
+                } catch (TransitionFatalException $e) {
+                    
+                } catch (TransitionRecoverableException $e) {
+                    // Left only recoverable problems
+                    return true;
+                }
+                return false;
+            })
+            ->map(function(\Closure $condition) {
+                try {
+                    call_user_func($condition, $this->model);
+                } catch (TransitionRecoverableException $e) {
+                    return $e->getMessage();
+                }
+            })
+            ->toArray();
     }
 
     /**
-     * Get attributes, that must be provided into transit() method
+     * Get attributes, that must be provided into transit() method.
+     *
      * @return Collection
      */
-    public function getRequirements()
+    public function getRequirements(): Collection
     {
         return $this->attributes;
     }
 
     /**
-     * Parent context of this transition
+     * Parent context of this transition.
+     *
      * @return StateMachineEngine|null
      */
-    protected function workflow()
+    protected function workflow(): ?StateMachineEngine
     {
         return $this->model->workflow($this->attribute);
     }
 
     /**
-     * Examine transition preconditions
+     * Examine transition preconditions.
+     *
      * @throws TransitionRecoverableException
      * @throws TransitionFatalException
      */
