@@ -1,41 +1,43 @@
 <?php
 
+namespace App;
 
+use Codewiser\Workflow\Exceptions\TransitionFatalException;
 use Codewiser\Workflow\Exceptions\TransitionRecoverableException;
 use Codewiser\Workflow\Transition;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\Notification;
 
 class Blueprint extends \Codewiser\Workflow\WorkflowBlueprint
 {
-
-    /**
-     * Array of available Model Workflow steps. First one is initial
-     * @return array|string[]
-     * @example [new, review, published, correcting]
-     */
     protected function states(): array
     {
-        return ['new', 'old'];
+        return ['one', 'two', 'three', 'four'];
     }
 
-    /**
-     * Array of allowed transitions between states
-     * @return array|Transition
-     * @example [[new, review], [review, published], [review, correcting], [correcting, review]]
-     */
     protected function transitions(): array
     {
         return [
-            Transition::define('new', 'old')
-                ->condition(function (Model $model) {
-                    if (strlen($model->body) < 1000) {
-                        throw new TransitionRecoverableException('Post body is too small. At least 1000 symbols required');
-                    }
+            Transition::define('one', 'two')
+                ->condition(function (Post $model) {
+                    throw new TransitionRecoverableException('Recoverable');
                 })
-                ->condition([$this, 'states'])
-                ->callback(function (Model $model, $payload) {
-                    $model->user->notify(new Notification($model, $payload['attr']));
-                })
+                ->authorize('publish'),
+
+            Transition::define('one', 'three')
+                ->condition(function (Post $model) {
+                    throw new TransitionFatalException('Fatal');
+                }),
+
+            Transition::define('one', 'four')
+                ->requires('comment')
+                ->callback(function (Post $model, $payload) {
+                    $model->body = 'sent';
+                }),
+
+            Transition::define('four', 'three'),
+            Transition::define('two', 'three'),
+            Transition::define('three', 'one')
         ];
     }
 }
