@@ -3,10 +3,11 @@
 namespace Codewiser\Workflow;
 
 use Codewiser\Workflow\Exceptions\WorkflowException;
-use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Support\Collection;
 
 /**
- * Workflow (aka State Machine) blueprint
+ * Workflow (aka State Machine) blueprint.
+ *
  * @package Codewiser\Workflow
  */
 abstract class WorkflowBlueprint
@@ -17,27 +18,31 @@ abstract class WorkflowBlueprint
     }
 
     /**
-     * Array of available Model Workflow steps. First one is initial
-     * @return array|string[]
+     * Array of available Model Workflow steps. First one is initial.
+     *
+     * @return string[]
      * @example [new, review, published, correcting]
      */
     abstract protected function states(): array;
 
     /**
-     * Array of allowed transitions between states
-     * @return array|Transition[]
+     * Array of allowed transitions between states.
+     *
+     * @return Transition[]
      * @example [[new, review], [review, published], [review, correcting], [correcting, review]]
      */
     abstract protected function transitions(): array;
 
     /**
-     * Validates State Machine Blueprint
+     * Validates State Machine Blueprint.
+     *
      * @throws WorkflowException
      */
     protected function validate()
     {
         $states = $this->getStates();
         $transitions = collect();
+        
         foreach ($this->getTransitions() as $transition) {
             $s = $transition->getSource();
             $t = $transition->getTarget();
@@ -48,28 +53,30 @@ abstract class WorkflowBlueprint
             if (!$states->contains($t)) {
                 throw new WorkflowException("Buggy blueprint: transition to nowhere");
             }
-            if ($transitions->contains($s.$t)) {
+            if ($transitions->contains($s . $t)) {
                 throw new WorkflowException("Buggy blueprint: transition duplicate");
             }
-            $transitions->push($s.$t);
+            $transitions->push($s . $t);
         }
     }
 
     /**
-     * Array of states
+     * Array of states.
+     *
      * @return string[]|Collection
      */
-    public function getStates()
+    public function getStates(): Collection
     {
         return collect($this->states());
     }
 
     /**
-     * Array of transitions between states
-     * @return Transition[]|Collection
+     * Array of transitions between states.
+     *
+     * @return TransitionCollection
      */
-    public function getTransitions()
+    public function getTransitions(): TransitionCollection
     {
-        return collect($this->transitions());
+        return TransitionCollection::make($this->transitions());
     }
 }
